Zmiana struktury customer
