+++ apps/backend/src/api/admin/customer/[id]/activate/route.ts
