--- apps/backend/src/api/admin/customers/activate/query-config.ts
@@ -1,18 +1,0 @@
-/* Company Query Config */
-export const approvedFields = [
-    "email",
-    "approved",
-  ];
-  
-  export const approvedQueryConfig = {
-    list: {
-      defaults: approvedFields,
-      isList: true,
-    },
-    retrieve: {
-      defaults: approvedFields,
-      isList: false,
-    },
-  };
-  
-
