--- apps/backend/src/api/admin/customers/get-customers/query-config.ts
@@ -1,21 +1,0 @@
-/* Company Query Config */
-export const accountsFields = [
-    "email",
-    "approved",
-    "created_at",
-    "updated_at",
-    "deleted_at"
-  ];
-  
-  export const accountsQueryConfig = {
-    list: {
-      defaults: accountsFields,
-      isList: true,
-    },
-    retrieve: {
-      defaults: accountsFields,
-      isList: false,
-    },
-  };
-  
-
